--- doc/modules/io.rst
@@ -1,9 +1,0 @@
-io - File reading and writing
-=============================
-
-.. currentmodule:: hcipy.io
-
-.. automodapi:: hcipy.io
-    :no-inheritance-diagram:
-    :include-all-objects:
-    :no-heading:
