--- doc/modules/math_util.rst
@@ -1,9 +1,0 @@
-math_util - Math Utilities
-==========================
-
-.. currentmodule:: hcipy.math_util
-
-.. automodapi:: hcipy.math_util
-    :no-inheritance-diagram:
-    :include-all-objects:
-    :no-heading:
