--- doc/modules/statistics.rst
@@ -1,9 +1,0 @@
-statistics - Statistical Utilities
-==================================
-
-.. currentmodule:: hcipy.statistics
-
-.. automodapi:: hcipy.statistics
-    :no-inheritance-diagram:
-    :include-all-objects:
-    :no-heading:
