+++ doc/modules/util.rst
@@ -1,0 +1,9 @@
+util - Miscellaneous Utilities
+==============================
+
+.. currentmodule:: hcipy.util
+
+.. automodapi:: hcipy.util
+    :no-inheritance-diagram:
+    :include-all-objects:
+    :no-heading:
