+++ hcipy/__init__.py
@@ -5,13 +5,11 @@ from . import coronagraphy
 from . import field
 from . import fourier
 from . import interpolation
-from . import io
-from . import math_util
 from . import mode_basis
 from . import optics
 from . import plotting
 from . import propagation
-from . import statistics
+from . import util
 from . import wavefront_control
 from . import wavefront_sensing
 
@@ -23,14 +21,12 @@ from .coronagraphy import *
 from .field import *
 from .fourier import *
 from .interpolation import *
-from .io import *
-from .math_util import *
 from .metrics import *
 from .mode_basis import *
 from .optics import *
 from .plotting import *
 from .propagation import *
-from .statistics import *
+from .util import *
 from .wavefront_control import *
 from .wavefront_sensing import *
 
@@ -43,14 +39,12 @@ __all__.extend(coronagraphy.__all__)
 __all__.extend(field.__all__)
 __all__.extend(fourier.__all__)
 __all__.extend(interpolation.__all__)
-__all__.extend(io.__all__)
-__all__.extend(math_util.__all__)
 __all__.extend(metrics.__all__)
 __all__.extend(mode_basis.__all__)
 __all__.extend(optics.__all__)
 __all__.extend(plotting.__all__)
 __all__.extend(propagation.__all__)
-__all__.extend(statistics.__all__)
+__all__.extend(util.__all__)
 __all__.extend(wavefront_control.__all__)
 __all__.extend(wavefront_sensing.__all__)
