+++ hcipy/atmosphere/infinite_atmospheric_layer.py
@@ -1,7 +1,7 @@
 from __future__ import division
 
 from .atmospheric_model import AtmosphericLayer, phase_covariance_von_karman, fried_parameter_from_Cn_squared
-from ..statistics import SpectralNoiseFactoryMultiscale
+from ..util import SpectralNoiseFactoryMultiscale
 from ..field import Field, RegularCoords, UnstructuredCoords, CartesianGrid
 from .finite_atmospheric_layer import FiniteAtmosphericLayer
