--- hcipy/io/__init__.py
@@ -1,5 +1,0 @@
-__all__ = ['read_fits', 'write_fits']#, 'write_mode_basis', 'read_mode_basis']
-__all__ += ['list_all_files_with_extension']
-
-from .fits import *
-from .util import *
