--- hcipy/math_util/__init__.py
@@ -1,5 +1,0 @@
-__all__ = ['inverse_truncated', 'inverse_truncated_modal', 'inverse_tikhonov']
-__all__ += ['SVD']
-
-from .matrix_inversion import *
-from .singular_value_decomposition import *
