--- hcipy/statistics/__init__.py
@@ -1,8 +1,0 @@
-__all__ = ['SpectralNoiseFactory', 'SpectralNoise']
-__all__ += ['SpectralNoiseFactoryFFT', 'SpectralNoiseFFT']
-__all__ += ['SpectralNoiseFactoryMultiscale', 'SpectralNoiseMultiscale']
-__all__ += ['large_poisson']
-
-from .kalman_filter import *
-from .spectral_noise import *
-from .util import *
