+++ hcipy/util/io.py
