+++ hcipy/util/singular_value_decomposition.py
