+++ hcipy/util/spectral_noise.py
